Ugly fix for maven embedder problem.
Some plugins write directly to stdout instead of using the logger. Therefore we have to redirect stdout while maven is running to
get the output of the embedder and redirect it back afterwards. This is an ugly way of doing this, but the only one that the maven
developers propose.

Signed-off-by: Michael Handler <michael.handler@fullstop.at>