First experiments with clover
Be carefull with this commit. Currently the clover branch does NOT work,
because of the "problems" with jibx as described in the mailinglist. As
soon as jibx is removed the clover branch can be finished

OPENENGSB-6

Signed-off-by: Andreas Pieber <anpieber@gmail.com>
(cherry picked from commit 82a9265c74d83db32cb136f53a66d6f92cff65e6)

Signed-off-by: Andreas Pieber <anpieber@gmail.com>