refactor service descriptor localization.
This patch injects the locale and string lookup into the service
descriptor builder for removing boilerplate code. Previously
implementors of a service had to do 'strings.getString(foo, locale)' to
lookup localization values, now they simply provide the key for lookup.
Internal lookup for AttributeDefinition is missing and comes in a later
patch.