add helper script for connector project generation.
The Maven Archetype possibilities are too limited. Developers still
have to input to much boilerplate variables to generate an initial
connector project. This is errorprone and inconvenient. This helper
script only needs a few inputs from the user and does to guess most
of the variables based on location and previous input.