moved opencit into own code base
The problem is pax:provision. This scans all child projects recursive
and load them into pax:provison. This is simply not what we want.
Nevertheless, if we do not include opencit into the build chain
(add it to the root.pom) it will not be builded automatically,
loosing much of the advantages of CI

Signed-off-by: Andreas Pieber <anpieber@gmail.com>