moved cnofiguration files into own preparation folder and package them
Thie allows to provide unchanged configuration files which can be unpacked
and configured using simple maven commands (to be shown in the assembly
project). The final artifacts can be repacked and provided in the openengsb
application or its childs.

This commit as is is NOT functional. The configuration files have been moved
and will (therefore) neither be available for the pax:provision command nor
for the release scenario. This have to be finished first.

Signed-off-by: Andreas Pieber <anpieber@gmail.com>