[OPENENGSB-1216] Add a general testing feature for OpenEngSBCoreServices
In the previous method is was almost impossible to develop testdriven
against the OpenEngSBCoreService and OsgiUtilServices. This is now
corrected by making a service out of the OsgiUtilServices and importing
them via the OpenEngsBCoreService. If you set the OsgiUtilService in the
OpenEngSBCoreService you can easily test it. A valuable general helper
is the AbstractOsgiMockServiceTest in the core-test package providing
most required methods for testing those services out of the box.

Signed-off-by: Andreas Pieber <anpieber@gmail.com>