[OPENENGSB-2804] Remove no longer required .gitmodules file
The .gitmodules file was introduced as we've used git module to get
access to all the openengsb modules at once. Since we use repo now to
achive this goal this file is no longer required and removed by this
patch to avoid the confusion when ppl might see that file in the src
releases.

Signed-off-by: Andreas Pieber <anpieber@gmail.com>